Trait communication_protocols::simulation_model::SimulationModel

pub trait SimulationModel: ErrorHandling {
    type Type;
    fn new(
        model: &SimpleModel, 
        state: &mut SimulationStateHeader, 
        n: usize
    ) -> Result<Self::Type, String>;

    fn march(
        &self, 
        date: Date, 
        weather: &dyn Weather, 
        model: &SimpleModel, 
        state: &mut SimulationState
    ) -> Result<(), String>;
}

Protocols for SimulationModels

Associated Types

type Type

Required methods

fn new(
model: &SimpleModel,
state: &mut SimulationStateHeader,
n: usize
) -> Result<Self::Type, String>

Creates a new Simulation Model

Arguments;

A model model

A model state model (data important to the Sim Engine will be added)

The number of timesteps per hour for the main simulation… internally, each simulation model will choose its own sub-subdivition (e.g. half or one third of what is asked for)

fn march(
&self,
date: Date,
weather: &dyn Weather,
model: &SimpleModel,
state: &mut SimulationState
) -> Result<(), String>

Marchs forward in the simulation.

Arguments

The model

The model state (will be modified)

A piece of weather data